// blockchain/typechain-types/PublicationRegistry.ts
export declare namespace PublicationRegistry {
  export type PublicationStruct = {
    author: AddressLike;
    timestamp: BigNumberish;
    exists: boolean;
  };

  export type PublicationStructOutput = [
    author: string,
    timestamp: bigint,
    exists: boolean
  ] & { author: string; timestamp: bigint; exists: boolean };
}

export interface PublicationRegistryInterface extends Interface {
  getFunction(
    nameOrSignature: "getPublication" | "publications" | "registerPublication"
  ): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "PublicationRegistered"): EventFragment;

  encodeFunctionData(
    functionFragment: "getPublication",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "publications",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "registerPublication",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "getPublication",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "publications",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "registerPublication",
    data: BytesLike
  ): Result;
}

export namespace PublicationRegisteredEvent {
  export type InputTuple = [
    contentHash: BytesLike,
    author: AddressLike,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    contentHash: string,
    author: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    contentHash: string;
    author: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface PublicationRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): PublicationRegistry;
  waitForDeployment(): Promise<this>;

  interface: PublicationRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getPublication: TypedContractMethod<
    [contentHash: BytesLike],
    [PublicationRegistry.PublicationStructOutput],
    "view"
  >;

  publications: TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, boolean] & {
        author: string;
        timestamp: bigint;
        exists: boolean;
      }
    ],
    "view"
  >;

  registerPublication: TypedContractMethod<
    [contentHash: BytesLike],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getPublication"
  ): TypedContractMethod<
    [contentHash: BytesLike],
    [PublicationRegistry.PublicationStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "publications"
  ): TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, bigint, boolean] & {
        author: string;
        timestamp: bigint;
        exists: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "registerPublication"
  ): TypedContractMethod<[contentHash: BytesLike], [void], "nonpayable">;

  getEvent(
    key: "PublicationRegistered"
  ): TypedContractEvent<
    PublicationRegisteredEvent.InputTuple,
    PublicationRegisteredEvent.OutputTuple,
    PublicationRegisteredEvent.OutputObject
  >;

  filters: {
    "PublicationRegistered(bytes32,address,uint256)": TypedContractEvent<
      PublicationRegisteredEvent.InputTuple,
      PublicationRegisteredEvent.OutputTuple,
      PublicationRegisteredEvent.OutputObject
    >;
    PublicationRegistered: TypedContractEvent<
      PublicationRegisteredEvent.InputTuple,
      PublicationRegisteredEvent.OutputTuple,
      PublicationRegisteredEvent.OutputObject
    >;
  };
}